requirements.txt
Flask>=2.0
python-pptx==0.6.23
Werkzeug>=2.0
openai>=1.0   # For OpenAI API calls
google-generativeai>=0.5 # For Gemini API calls
# libreoffice (or soffice command) needs to be installed on the system for PDF conversion, webapp will still work o/w
requests
pandas
tqdm
Pillow==10.4.0
pdf2image==1.17.0
opencv-python==4.9.0.80
numpy